backend/requirements.txt
# AcuRate Backend Requirements
# Python 3.12+

# Django Framework
# SECURITY: Updated from 5.2.1 to fix CVE-2025-* vulnerabilities
Django>=5.2.9

# Django REST Framework
djangorestframework==3.15.2

# PostgreSQL Database Adapter
psycopg2-binary==2.9.10

# JWT Authentication
# SECURITY: Updated from 5.3.1 to fix CVE-2024-22513
djangorestframework-simplejwt>=5.5.1

# CORS Headers
django-cors-headers==4.6.0

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.1

# Image Handling
Pillow==11.0.0

# Date Utilities
python-dateutil==2.9.0

# Email / SendGrid
django-sendgrid-v5==1.3.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
coverage==7.6.1

# API Documentation
drf-spectacular==0.29.0

# Rate Limiting
django-ratelimit==4.1.0

# Logging
python-json-logger==2.0.7

# Redis Cache (Optional - for production caching)
django-redis==5.4.0

# Background Tasks (Optional - for async operations)
celery==5.3.4
celery[redis]==5.3.4

# Error Tracking (Optional - for production monitoring)
sentry-sdk==2.15.0

# Production Server
gunicorn==23.0.0

# Static Files (Optional - for production static file serving)
whitenoise==6.8.2

